--- src/main/java/com/example/mybootapp3/Mybootapp3Application.java
@@ -1,13 +1,0 @@
-package com.example.mybootapp3;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Mybootapp3Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Mybootapp3Application.class, args);
-	}
-
-}
