+++ src/main/java/jp/te4a/spring/boot/myapp5/mybootapp5/Mybootapp3Application.java
@@ -1,0 +1,13 @@
+package jp.te4a.spring.boot.myapp5.mybootapp5;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Mybootapp3Application {
+
+	public static void main(String[] args) {
+		SpringApplication.run(Mybootapp3Application.class, args);
+	}
+
+}
