+++ src/main/resources/application.properties
@@ -1,1 +1,3 @@
 spring.application.name=mybootapp3
+spring.mvc.view.prefix=/WEB-INF/views/
+spring.mvc.view.suffix=.jsp
