--- src/test/java/com/example/mybootapp3/Mybootapp3ApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.mybootapp3;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Mybootapp3ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
